Continuation of auth enhancements

+++ server/modules/filedatastore/filedatastoreimpl.go
@@ -11,41 +11,46 @@
 package filedatastore
 
 import (
+  "context"
   "errors"
   "fmt"
+  "github.com/apex/log"
+  "github.com/kennygrant/sanitize"
+  "github.com/security-onion-solutions/securityonion-soc/json"
+  "github.com/security-onion-solutions/securityonion-soc/model"
+  "github.com/security-onion-solutions/securityonion-soc/module"
+  "github.com/security-onion-solutions/securityonion-soc/packet"
+  "github.com/security-onion-solutions/securityonion-soc/server"
+  "github.com/security-onion-solutions/securityonion-soc/web"
   "io"
   "os"
   "path/filepath"
   "strings"
   "sync"
   "time"
-  "github.com/apex/log"
-  "github.com/security-onion-solutions/securityonion-soc/json"
-  "github.com/security-onion-solutions/securityonion-soc/model"
-  "github.com/security-onion-solutions/securityonion-soc/module"
-  "github.com/security-onion-solutions/securityonion-soc/packet"
-  "github.com/kennygrant/sanitize"
 )
 
 const DEFAULT_RETRY_FAILURE_INTERVAL_MS = 600000
 
 type FileDatastoreImpl struct {
-  jobDir                  string
-  retryFailureIntervalMs  int
-  jobsByNodeId            map[string][]*model.Job
-  jobsById                map[int]*model.Job
-  nodesById               map[string]*model.Node
-  ready                   bool
-  nextJobId               int
-  lock                    sync.RWMutex
+  server                 *server.Server
+  jobDir                 string
+  retryFailureIntervalMs int
+  jobsByNodeId           map[string][]*model.Job
+  jobsById               map[int]*model.Job
+  nodesById              map[string]*model.Node
+  ready                  bool
+  nextJobId              int
+  lock                   sync.RWMutex
 }
 
-func NewFileDatastoreImpl() *FileDatastoreImpl {
-  return &FileDatastoreImpl {
+func NewFileDatastoreImpl(srv *server.Server) *FileDatastoreImpl {
+  return &FileDatastoreImpl{
+    server:       srv,
     jobsByNodeId: make(map[string][]*model.Job),
-    jobsById: make(map[int]*model.Job),
-    nodesById: make(map[string]*model.Node),
-    lock: sync.RWMutex{},
+    jobsById:     make(map[int]*model.Job),
+    nodesById:    make(map[string]*model.Node),
+    lock:         sync.RWMutex{},
   }
 }
 
@@ -82,8 +87,8 @@ func (datastore *FileDatastoreImpl) AddNode(node *model.Node) error {
 
 func (datastore *FileDatastoreImpl) addNode(node *model.Node) *model.Node {
   datastore.nodesById[node.Id] = node
-  log.WithFields(log.Fields {
-    "id": node.Id,
+  log.WithFields(log.Fields{
+    "id":          node.Id,
     "description": node.Description,
   }).Debug("Added node")
   return node
@@ -101,15 +106,15 @@ func (datastore *FileDatastoreImpl) UpdateNode(newNode *model.Node) (*model.Node
     }
 
     // Only copy the following values from the incoming node. Preserve everything else.
-    node.EpochTime     = newNode.EpochTime
-    node.Role          = newNode.Role
-    node.Description   = newNode.Description
-    node.Address       = newNode.Address
-    node.Version       = newNode.Version
+    node.EpochTime = newNode.EpochTime
+    node.Role = newNode.Role
+    node.Description = newNode.Description
+    node.Address = newNode.Address
+    node.Version = newNode.Version
 
     // Ensure model parameters are updated
     node.SetModel(newNode.Model)
-    
+
     // Mark ConnectionStatus as Ok since this node just checked in
     node.ConnectionStatus = model.NodeStatusOk
 
@@ -122,73 +127,115 @@ func (datastore *FileDatastoreImpl) UpdateNode(newNode *model.Node) (*model.Node
     log.WithField("description", newNode.Description).Info("Not adding node with missing id")
   }
   return node, nil
-} 
+}
 
-func (datastore *FileDatastoreImpl) GetNextJob(nodeId string) *model.Job {
+func (datastore *FileDatastoreImpl) GetNextJob(ctx context.Context, nodeId string) *model.Job {
   datastore.lock.RLock()
   defer datastore.lock.RUnlock()
   var nextJob *model.Job
-  now := time.Now()
-  jobs := datastore.jobsByNodeId[strings.ToLower(nodeId)]
-  for _, job := range jobs {
-    retryTime := job.FailTime.Add(time.Millisecond * time.Duration(datastore.retryFailureIntervalMs))
-    if job.Status != model.JobStatusCompleted && 
-       (nextJob == nil || job.CreateTime.Before(nextJob.CreateTime)) && 
-       (job.Status != model.JobStatusIncomplete || retryTime.Before(now)) {
-      nextJob = job
+
+  if err := datastore.server.Authorizer.CheckContextOperationAuthorized(ctx, "process", "jobs"); err == nil {
+    now := time.Now()
+    jobs := datastore.jobsByNodeId[strings.ToLower(nodeId)]
+    for _, job := range jobs {
+      retryTime := job.FailTime.Add(time.Millisecond * time.Duration(datastore.retryFailureIntervalMs))
+      if job.Status != model.JobStatusCompleted &&
+        (nextJob == nil || job.CreateTime.Before(nextJob.CreateTime)) &&
+        (job.Status != model.JobStatusIncomplete || retryTime.Before(now)) {
+        nextJob = job
+      }
     }
   }
   return nextJob
 }
 
-func (datastore *FileDatastoreImpl) CreateJob() *model.Job {
+func (datastore *FileDatastoreImpl) CreateJob(ctx context.Context) *model.Job {
   datastore.lock.Lock()
   defer datastore.lock.Unlock()
   job := model.NewJob()
   job.Id = datastore.nextJobId
   datastore.incrementJobId(job.Id)
-  log.WithFields(log.Fields {
-    "id": job.Id,
+  log.WithFields(log.Fields{
+    "id":        job.Id,
     "nextJobId": datastore.nextJobId,
   }).Debug("Created job")
 
   return job
 }
 
-func (datastore *FileDatastoreImpl) GetJob(jobId int) *model.Job {
-  datastore.lock.RLock()
-  defer datastore.lock.RUnlock()
-  return datastore.getJobById(jobId)
+func (datastore *FileDatastoreImpl) jobIsAllowed(ctx context.Context, job *model.Job, op string) bool {
+  allowed := false
+
+  if job != nil {
+    if err := datastore.server.Authorizer.CheckContextOperationAuthorized(ctx, op, "jobs"); err == nil {
+      // User can operate on all jobs
+      allowed = true
+    } else {
+      // User is only authorized against their own jobs.
+      if user, ok := ctx.Value(web.ContextKeyRequestor).(*model.User); ok {
+        if job.UserId == user.Id {
+          allowed = true
+        }
+      }
+    }
+  }
+  return allowed
 }
 
-func (datastore *FileDatastoreImpl) GetJobs() []*model.Job {
+func (datastore *FileDatastoreImpl) GetJob(ctx context.Context, jobId int) *model.Job {
+  datastore.lock.RLock()
+  defer datastore.lock.RUnlock()
+  job := datastore.getJobById(jobId)
+  if job != nil {
+    if !datastore.jobIsAllowed(ctx, job, "read") {
+      // Do not return jobs that are not allowed to be viewed by this user.
+      job = nil
+    }
+  }
+
+  return job
+}
+
+func (datastore *FileDatastoreImpl) GetJobs(ctx context.Context) []*model.Job {
   datastore.lock.RLock()
   defer datastore.lock.RUnlock()
   allJobs := make([]*model.Job, 0)
   for _, job := range datastore.jobsById {
-    allJobs = append(allJobs, job)
+    if datastore.jobIsAllowed(ctx, job, "read") {
+      allJobs = append(allJobs, job)
+    }
   }
   return allJobs
 }
 
-func (datastore *FileDatastoreImpl) AddJob(job *model.Job) error {
-  datastore.lock.Lock()
-  defer datastore.lock.Unlock()
-  err := datastore.addJob(job)
-  if err == nil {
-    err = datastore.saveJob(job)
+func (datastore *FileDatastoreImpl) AddJob(ctx context.Context, job *model.Job) error {
+  var err error
+  if err = datastore.server.Authorizer.CheckContextOperationAuthorized(ctx, "write", "jobs"); err == nil {
+    datastore.lock.Lock()
+    defer datastore.lock.Unlock()
+    err = datastore.addJob(job)
+    if err == nil {
+      err = datastore.saveJob(job)
+    }
   }
   return err
 }
 
-func (datastore *FileDatastoreImpl) UpdateJob(job *model.Job) error {
-  datastore.lock.Lock()
-  defer datastore.lock.Unlock()
-  err := datastore.deleteJob(job)
-  if err == nil {
-    err = datastore.addJob(job)
-    if err == nil {
-      err = datastore.saveJob(job)
+func (datastore *FileDatastoreImpl) UpdateJob(ctx context.Context, job *model.Job) error {
+  var err error
+  if err = datastore.server.Authorizer.CheckContextOperationAuthorized(ctx, "process", "jobs"); err == nil {
+    if job.CanProcess() {
+      datastore.lock.Lock()
+      defer datastore.lock.Unlock()
+      err = datastore.deleteJob(job)
+      if err == nil {
+        err = datastore.addJob(job)
+        if err == nil {
+          err = datastore.saveJob(job)
+        }
+      }
+    } else {
+      err = errors.New("Job is ineligible for processing")
     }
   }
   return err
@@ -198,24 +245,27 @@ func (datastore *FileDatastoreImpl) getJobById(jobId int) *model.Job {
   return datastore.jobsById[jobId]
 }
 
-func (datastore *FileDatastoreImpl) DeleteJob(job *model.Job) error {
-  err := datastore.deleteJob(job)
-  if err == nil {
-    job.Status = model.JobStatusDeleted
-    filename := fmt.Sprintf("%d.json", job.Id)
-    folder := filepath.Join(datastore.jobDir, sanitize.Name(job.GetNodeId()))
-    err = os.Remove(filepath.Join(folder, filename))
+func (datastore *FileDatastoreImpl) DeleteJob(ctx context.Context, job *model.Job) error {
+  var err error
+  if datastore.jobIsAllowed(ctx, job, "delete") {
+    err = datastore.deleteJob(job)
     if err == nil {
-      filename = fmt.Sprintf("%d.bin", job.Id)
-      os.Remove(filepath.Join(folder, filename))
-      filename = fmt.Sprintf("%d.bin.unwrapped", job.Id)
-      os.Remove(filepath.Join(folder, filename))
+      job.Status = model.JobStatusDeleted
+      filename := fmt.Sprintf("%d.json", job.Id)
+      folder := filepath.Join(datastore.jobDir, sanitize.Name(job.GetNodeId()))
+      err = os.Remove(filepath.Join(folder, filename))
+      if err == nil {
+        filename = fmt.Sprintf("%d.bin", job.Id)
+        os.Remove(filepath.Join(folder, filename))
+        filename = fmt.Sprintf("%d.bin.unwrapped", job.Id)
+        os.Remove(filepath.Join(folder, filename))
 
-      log.WithFields(log.Fields {
-        "id": job.Id,
-        "folder": folder,
-        "filename": filename,
-      }).Info("Permanently deleted job and job files")
+        log.WithFields(log.Fields{
+          "id":       job.Id,
+          "folder":   folder,
+          "filename": filename,
+        }).Info("Permanently deleted job and job files")
+      }
     }
   }
   return err
@@ -236,8 +286,8 @@ func (datastore *FileDatastoreImpl) deleteJob(job *model.Job) error {
     }
     datastore.jobsByNodeId[job.GetNodeId()] = newJobs
     delete(datastore.jobsById, job.Id)
-    log.WithFields(log.Fields {
-      "id": job.Id,
+    log.WithFields(log.Fields{
+      "id":   job.Id,
       "node": job.GetNodeId(),
     }).Debug("Deleted job from list")
   }
@@ -257,8 +307,8 @@ func (datastore *FileDatastoreImpl) addJob(job *model.Job) error {
     datastore.jobsByNodeId[job.GetNodeId()] = append(jobs, job)
     datastore.jobsById[job.Id] = job
     datastore.incrementJobId(job.Id)
-    log.WithFields(log.Fields {
-      "id": job.Id,
+    log.WithFields(log.Fields{
+      "id":   job.Id,
       "node": job.GetNodeId(),
     }).Debug("Added job")
   }
@@ -274,9 +324,9 @@ func (datastore *FileDatastoreImpl) incrementJobId(id int) {
 func (datastore *FileDatastoreImpl) saveJob(job *model.Job) error {
   filename := fmt.Sprintf("%d.json", job.Id)
   folder := filepath.Join(datastore.jobDir, sanitize.Name(job.GetNodeId()))
-  log.WithFields(log.Fields {
-    "id": job.Id,
-    "folder": folder,
+  log.WithFields(log.Fields{
+    "id":       job.Id,
+    "folder":   folder,
     "filename": filename,
   }).Debug("Saving job file")
   os.MkdirAll(folder, os.ModePerm)
@@ -308,17 +358,21 @@ func (datastore *FileDatastoreImpl) loadJobs() error {
   return err
 }
 
-func (datastore *FileDatastoreImpl) GetPackets(jobId int, offset int, count int, unwrap bool) ([]*model.Packet, error) {
+func (datastore *FileDatastoreImpl) GetPackets(ctx context.Context, jobId int, offset int, count int, unwrap bool) ([]*model.Packet, error) {
   var packets []*model.Packet
   var err error
-  job := datastore.GetJob(jobId)
+  job := datastore.GetJob(ctx, jobId)
   if job != nil {
-    if job.Status == model.JobStatusCompleted {
-      packets, err = packet.ParsePcap(datastore.getStreamFilename(job), offset, count, unwrap)
-      if err != nil {
-        log.WithError(err).WithField("jobId", job.Id).Warn("Failed to parse captured packets")
-        err = nil
+    if datastore.jobIsAllowed(ctx, job, "read") {
+      if job.Status == model.JobStatusCompleted {
+        packets, err = packet.ParsePcap(datastore.getStreamFilename(job), offset, count, unwrap)
+        if err != nil {
+          log.WithError(err).WithField("jobId", job.Id).Warn("Failed to parse captured packets")
+          err = nil
+        }
       }
+    } else {
+      err = errors.New("Job is inaccessible")
     }
   } else {
     err = errors.New("Job not found")
@@ -327,55 +381,65 @@ func (datastore *FileDatastoreImpl) GetPackets(jobId int, offset int, count int,
   return packets, err
 }
 
-func (datastore *FileDatastoreImpl) SavePacketStream(jobId int, reader io.ReadCloser) error {
+func (datastore *FileDatastoreImpl) SavePacketStream(ctx context.Context, jobId int, reader io.ReadCloser) error {
   var err error
-  job := datastore.GetJob(jobId)
-  if job != nil {
-    var count int64
-    file, err := os.Create(datastore.getStreamFilename(job))
-    if err == nil {
-      defer file.Close()
-      count, err = io.Copy(file, reader)
-    }
-    if err != nil {
-      log.WithError(err).WithField("jobId", jobId).Error("Failed to write packet stream to file")
+  if err = datastore.server.Authorizer.CheckContextOperationAuthorized(ctx, "process", "jobs"); err == nil {
+    job := datastore.GetJob(ctx, jobId)
+    if job != nil {
+      if job.CanProcess() {
+        var count int64
+        file, err := os.Create(datastore.getStreamFilename(job))
+        if err == nil {
+          defer file.Close()
+          count, err = io.Copy(file, reader)
+        }
+        if err != nil {
+          log.WithError(err).WithField("jobId", jobId).Error("Failed to write packet stream to file")
+        } else {
+          log.WithFields(log.Fields{
+            "bytes": count,
+            "jobId": jobId,
+          }).Info("Saved packet stream to file")
+        }
+      } else {
+        err = errors.New("Job is ineligible for processing")
+      }
     } else {
-      log.WithFields(log.Fields {
-        "bytes": count,
-        "jobId": jobId,
-      }).Info("Saved packet stream to file")
+      err = errors.New("Job not found")
     }
-  } else {
-    err = errors.New("Job not found")
   }
   return err
 }
 
-func (datastore *FileDatastoreImpl) GetPacketStream(jobId int, unwrap bool) (io.ReadCloser, string, int64, error) {
+func (datastore *FileDatastoreImpl) GetPacketStream(ctx context.Context, jobId int, unwrap bool) (io.ReadCloser, string, int64, error) {
   var reader io.ReadCloser
   var filename string
   var length int64
   var err error
-  job := datastore.GetJob(jobId)
+  job := datastore.GetJob(ctx, jobId)
   if job != nil {
-    if job.Status == model.JobStatusCompleted {
-      filename = fmt.Sprintf("sensoroni_%s_%d.%s", sanitize.Name(job.GetNodeId()), job.Id, sanitize.Name(job.FileExtension));
-      var file *os.File
-      file, err = os.Open(datastore.getModifiedStreamFilename(job, unwrap))
-      if err != nil {
-        log.WithError(err).WithField("jobId", job.Id).Error("Failed to open packet stream")
-      } else {
-        reader = file
-        info, err := file.Stat()
-        length = info.Size()
-        log.WithFields(log.Fields {
-          "size": length,
-          "name": info.Name(),
-        }).Info("Streaming file")
+    if datastore.jobIsAllowed(ctx, job, "read") {
+      if job.Status == model.JobStatusCompleted {
+        filename = fmt.Sprintf("sensoroni_%s_%d.%s", sanitize.Name(job.GetNodeId()), job.Id, sanitize.Name(job.FileExtension))
+        var file *os.File
+        file, err = os.Open(datastore.getModifiedStreamFilename(job, unwrap))
         if err != nil {
-          log.WithError(err).WithField("jobId", job.Id).Error("Failed to open file stats")
+          log.WithError(err).WithField("jobId", job.Id).Error("Failed to open packet stream")
+        } else {
+          reader = file
+          info, err := file.Stat()
+          length = info.Size()
+          log.WithFields(log.Fields{
+            "size": length,
+            "name": info.Name(),
+          }).Info("Streaming file")
+          if err != nil {
+            log.WithError(err).WithField("jobId", job.Id).Error("Failed to open file stats")
+          }
         }
       }
+    } else {
+      err = errors.New("Job is inaccessible")
     }
   } else {
     err = errors.New("Job not found")
